Fixed missing file duration in Details Page

+++ TMessagesProj/src/main/java/it/foxgram/ui/DetailsActivity.java
@@ -397,8 +397,8 @@ public class DetailsActivity extends BaseFragment implements NotificationCenter.
                     for (int a = 0; a < messageObject.messageOwner.media.document.attributes.size(); a++) {
                         if (messageObject.messageOwner.media.document.attributes.get(a) instanceof TLRPC.TL_documentAttributeAudio ||
                                 messageObject.messageOwner.media.document.attributes.get(a) instanceof TLRPC.TL_documentAttributeVideo) {
-                            int duration = (int) messageObject.messageOwner.media.document.attributes.get(a).duration;
-                            durationString = AndroidUtilities.formatShortDuration(duration);
+                            double duration = messageObject.messageOwner.media.document.attributes.get(a).duration;
+                            durationString = AndroidUtilities.formatShortDuration((int) duration);
                             fileDuration = rowCount++;
                         } else if (messageObject.messageOwner.media.document.attributes.get(a) instanceof TLRPC.TL_documentAttributeSticker) {
                             TLRPC.TL_documentAttributeSticker attribute = (TLRPC.TL_documentAttributeSticker) messageObject.messageOwner.media.document.attributes.get(a);
